Handle remote system encoding
Input command, stdout and stderr should be encoded/decoded with the
remote system encoding.

The encoding detection is lazy, it occur only when stdout/stderr are
requested and contains non ascii characters.

The encoding detection require python to be installed on remote system,
otherwise testinfra will assume the remote encoding is the same as the
local encoding.