Service: fix Upstart is_running test
When a service is managed by Upstart, the "service <my-service> status"
command returns exit code 0 even when the service is stopped, which
means UpstartService.is_running was true even when the service wasn't
running.

So instead of relying on the exit code, match the "running" string in
the output of "status <my-service>".