single fixture 'host'
Implement a new fixture 'host' that provide access to all modules though
attributes in non camel case names (eg. system_info for SystemInfo).

This is a major enhancement for readability of the code.

Provide backward compatibility (legacy fixtures) until version 2.0
This enhancement require a huge refactoring of the documentation and
some drops ("make your own module" is to re-think now...).

Closes #151