Fix --nagios option with py.test and deprecate testinfra command
When using testinfra command, an anoying warning "Module already
imported so can not be re-written: testinfra" is displayed.

The only reason of the existence of testinfra command as a wrapper for
pytest was handling '--nagios' option that capture py.test output and
display nagios plugin compatible output *before* pytest test report.

This is now fixed by switching the output file of pytest's
terminalreporter to a in-memory temporary file.

Since the --nagios option work with py.test command, deprecate testinfra
command since it become useless and generate an anoying warning.

Closes #135