Fix for 'linked_to' test for BSD-like systems
While running testinfra against a Mac OS system, I noticed that
the `linked_to` test uses `readlink -f` to evaluate a canonicalized
real path by recursively following all symlinks. Though this behavior
works with GNU's `readlink`, it does not work the same way for BSD's
version. In fact, it errors out because BSD's `readlink` does not
understand the `-f` flag.

It is possible to install the GNU equivalent for `readlink` on MacOS
via Homebrew. However, it seems like poor design to force this method
to have a dependency on Homebrew.

I found a StackOverflow post showing how to recreate the behavior of
GNU's `readlink` using BSD's `readlink`. This change is based on the
solution provided in that post, which fixes the behavior of the
`linked_to` test for systems using the BSD version of `readlink`
(e.g. MacOS). It should behave just like `readlink -f` on Linux systems.

Reference: https://stackoverflow.com/questions/1055671/how-can-i-get-the-behavior-of-gnus-readlink-f-on-a-mac