Use the RequireTTY setting if specified in a provided SSHConfig
Currently the ssh_config option RequireTTY is when using the
paramiko backend with an ssh config file.  This option is useful
when using sudo and connecting to hosts which have the requiretty
option set in sshd_config. Setting requiretty in sshd_config is no
longer a common default, but it is frequently on older distributions and
in some 'hardened' distributions.

This change causes paramiko to use Channel.get_pty() to request a
pseudo-terminal if the RequireTTY option is specified in a provided
ssh_config.

This allows addressing

    https://github.com/philpep/testinfra/issues/247

without modifying the server-side configuration.