Add namespace support for kubectl connector. (#317)
Add a `namespace` query string argument, so it's not necessary to switch using kubectl to work with a specific namespace.
Also fix specifying container in the pod, the feature using `/` was broken because urllib split the url into `url` and `path`, so this has never worked. Add a query string argument `container` instead.