Add '-w' argument to iptables
If multiple tests are executed in parallel (e.g. via pytest-xdist)
and more than one test queries the iptables rules on the same host
simultaneously, the following error will occur:

    E       AssertionError: Unexpected exit code 4 for
    CommandResult(command='iptables -t filter -S', exit_status=4,
    stdout='', stderr='Another app is currently holding the xtables
    lock. Perhaps you want to use the -w option?')

Iptables has an internal lock to prevent multiple simultaneous
invocations.  This change implements that suggestion and adds "-w 90"
to the iptables command so that we will wait up to 90 seconds to
obtain the lock.