test: fix pytest 4 markers usage
A getattr on the function doesn't work since pytest 4.
Looks the correct way to do this is to test the presence of "pytestmark"
attribute on the function which is either undefined or contains a list
of Marker objects.
We don't care about backward compat with older pytest versions for
testinfra own tests.