Add basic support for Windows users
Some options like `home` are not easily captured for a user other than
the user running the command. The `password` option is not possible due
to the way Windows handles the SAM file. Other options like `shell`
don't make sense in the Windows context and have been marked as such by
raising `NotImplementedError`.