Handle literal ipv6 addresses
An ansible inventory might specify the "ansible_host" as a literal
ipv6 address.

This needs handling in ingestion; in ansible_runner.py we add a check
if the host looks like a literal ipv6 address and enclose it in []'s
per RFC2732 (a util function is added for this, which works better on
Python 3).

When connecting, testinfra's host parsing in backend/base.py then
needs to check for []'s in the connection string and extract the
literal ipv6 address if found.  The test-case for this is updated with
a few ipv6 examples.