Deploy pypi releases with travis
See documentation on https://docs.travis-ci.com/user/deployment/pypi/

Add a "deploy" stage after lint and tests that is triggered only for tag
releases and only if the repo is "philpep/testinfra" to avoid triggering
this stage for forked repos.

We use the new token api for pypi, the password is a secure environment
variable PYPI_PASSWORD only available on master branch builds.

Upload sdist and bdist_wheel.