Fix safe-ssh backend with shell redirections
safe-ssh already use a lot of redirection and was expecting the input
command was a single command with arguments.

When testing with more complex commands, the code may crash.
Avoid this by wrapping the command in a "sh -c" execution.

Add a sanity check for "host.run()" testing exit code, stdout and stderr
are decoded correctly for all tested backends (ssh, safe-ssh, paramiko,
ansible, ansible with force_ansible etc).