Fix pure ansible connections with ANSIBLE_ANY_ERRORS_FATAL=True
When ANSIBLE_ANY_ERRORS_FATAL is set to True (this can be set from
environment, and probably from inventory or ansible configuration
files), the ansible call exit with status 8 when running a command that
fail to run.

Detail of exit codes can be found here:
https://github.com/ansible/ansible/blob/523e40e993c06a6c15234c9a3c951b5398e44184/lib/ansible/executor/task_queue_manager.py#L63-L68

Add `8` to the expected exit codes and add a non regression test.

Closes #481