Add extras for backend dependencies (#454)
* Add extras for backend dependencies

This adds environment selectors for backends in setup.cfg.  The
inspiration here is that the ansible backend actually requires
paramiko but it isn't specified anywhere.  Having environment markers
makes it a bit eaiser in CI environments to ensure everything is
present.

I've taken a guess at the requirements for the others; although
they're simple I think it's good for consistency to have all backends
with their own environment.  A documentation note is added too.