Fix module class for RPM and Debian distributions #570
A CentOS system with dpkg package installed fails all packages
is_installed checks. This happens because the decision to use module class
is based on the presence of `dpkg-query` binary.

This commit change the check to look for `system_info.distribution`, and
makes CentOS 6 distribution name lower case to be consistent.