backend: Deduplicate hosts returned by get_backends()
If a host appears multiple times in host specs (which can typically
happen if a host is a member of multiple Ansible groups for instance),
testinfra would run tests on that host multiple times.

If a host finds itself listed multiple times with the same backend and
connection parameters, testinfra will run the tests on that host only
once.

Signed-off-by: Benoît Knecht <bknecht@protonmail.ch>
Co-Authored-By: Philippe Pepiot <phil@philpep.org>