Updating uname check
Reverted back to the original code and checking for Windows
via the current `uname` implementation. Normally it will give you
back a `uname.rc == 1` with the exception if you have any GNU tools
installed on Windows. Therefore I currently check just for `msys` in the
beginning. I assume there is no Linux distribution that begins with
this 4 letters.