Improved RpmPackage is_installed property logic.
This new logic verifies if there is a "not installed" string in
the stdout when "rpm -q %s" command returns 1.

As seen in bug #758, if rpm command failed because of other reasons (such
as database corruption), the old implementation was returning True,
making the end-user think that the package was installed, however it was
never queried.